made swap separate section inside trade screen
