+++ src/screens/trade/children/index.ts
@@ -1,3 +1,4 @@
+export * from './swapTokenContent'
 export * from './swapAmountInput';
 export * from './swapFeeSection';
 export * from './errorSection';
